caculate label data

+++ segmentation/dataset/LungDataset.py
@@ -137,9 +137,8 @@ class LungDataset(Dataset):
         for i in range(batch_size):
             img = conver_numpy_image(list_imgs[i])
             mask = conver_numpy_image(list_masks[i])
-            # import pdb; pdb.set_trace()
-            img = (img +1)*255
-            img = img.long()
+            import pdb; pdb.set_trace()
+            img = int((img +1)*255)
             ct = contour(img, mask)
             if img.shape[2]!=mask.shape[2]:
                mask = np.concatenate([mask]*3, axis = 2)*255
